fix Appointment, refact: route feat: request

+++ app/Http/Controllers/Service/ServiceController.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace App\Http\Controllers\Service;
 
 use App\Http\Controllers\Controller;
@@ -6,42 +7,54 @@ use App\Http\Requests\Service\ServiceRequest;
 use App\Services\ServiceService;
 use App\Models\Service;
 use Illuminate\Container\Attributes\Auth;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
 class ServiceController extends Controller
 {
     public function __construct(
         protected ServiceService $serviceService
-    ) {
-        $this->authorizeResource(Service::class, 'service');
-    }
+    ) {}
 
-    public function index(ServiceRequest $request)
+    public function index()
     {
+        $this->authorize('viewAny', Service::class);
 
         $services = $this->serviceService->listServices();
 
         return response()->json($services);
     }
 
-    public function store(ServiceRequest $request):Response
+    public function show(Service $service)
     {
-        $this->serviceService->createService($request->validated());
+        $this->authorize('view', $service);
+        $services = $this->serviceService->findService($service);
 
-        return response()->json(null,Response::HTTP_CREATED);
+        return response()->json($services);
     }
 
-
-
-    public function update(ServiceRequest $request, int $id): Response
+    public function store(ServiceRequest $request): Response
     {
-        $this->serviceService->updateService($id, $request->validated());
+        $this->authorize('create', Service::class);
+
+        $this->serviceService->createService($request->validated());
+
+        return response()->json(null, Response::HTTP_CREATED);
+    }
+
+    public function update(ServiceRequest $request, Service $service): Response
+    {
+        $this->authorize('update', $service);
+
+        $updatedService = $this->serviceService->updateService($service->id, $request->validated());
 
         return response()->json(Response::HTTP_OK);
     }
 
-    public function destroy(Service $service):Response
+    public function destroy(Service $service): Response
     {
+        $this->authorize('update', $service);
+
 
         $service->delete();
         return response()->json(Response::HTTP_NO_CONTENT);
@@ -49,6 +62,8 @@ class ServiceController extends Controller
 
     public function restore(Service $service): Response
     {
+        $this->authorize('restore', $service);
+
         $service->restore();
 
         return response()->json(Response::HTTP_NO_CONTENT);
